Finished first pass of create-buildarchive.
It now creates the manifest.xml files in the appropriate build
archive directory.  It also now ignores any manifests that have "NO-BUILD"
in the manifest.xml.

It relies on EC to get the build number from the "Contrail-CI-CD" project.
If it tries to get the build number for a particular branch and it doesn't
exist, then it will create it with the initial build number of 100
and continue on.  [Presumably - Haven't actually tested that]