Had a bug when searching through the manifest file and grep for "NO-BUILD"
The logic deleted the entry if it found "NO-BUILD" but that would re-order the list
and the next entry would get skipped (3 becomes 2 and we already did 2 in the list).

Added another variable and make a copy fo the list and used that to search so
that wouldn't get modified during execution.